+++ R/drake_build.R
@@ -172,7 +172,7 @@ debug_command <- function(command) {
     debug_command_char(command)
   } else {
     . <- NULL
-    out <- safe_deparse(command)
+    out <- safe_deparse(command, backtick = TRUE)
     out <- debug_command_char(out)
     parse(text = out)[[1]]
   }