+++ R/process_imports.R
@@ -56,9 +56,9 @@ get_import_from_memory <- function(target, config) {
   lang <- parsed[[1]]
   is_namespaced <- length(lang) > 1
   if (is_namespaced) {
-    stopifnot(safe_deparse(lang[[1]]) %in% c("::", ":::"))
-    pkg <- safe_deparse(lang[[2]])
-    fun <- safe_deparse(lang[[3]])
+    stopifnot(safe_deparse(lang[[1]], backtick = FALSE) %in% c("::", ":::"))
+    pkg <- safe_deparse(lang[[2]], backtick = FALSE)
+    fun <- safe_deparse(lang[[3]], backtick = FALSE)
     tryCatch(get(fun, envir = getNamespace(pkg)), error = error_na)
   } else {
     NA_character_